split header to avoid depending on AABB-tree for basic orientation functions
